+++ internal/scanner/registry_default.go
@@ -3,11 +3,11 @@ package scanner
 // DefaultRegistry returns a registry with all MVP ecosystem scanners.
 func DefaultRegistry() *Registry {
 	reg := NewRegistry()
-	reg.Register(NewNodeScanner())
-	reg.Register(NewRustScanner())
-	reg.Register(NewRubyScanner())
-	reg.Register(NewPythonScanner())
-	reg.Register(NewGoScanner())
+	reg.Register(newWalkScanner(nodeWalkEcosystem))
+	reg.Register(newWalkScanner(rustWalkEcosystem))
+	reg.Register(newWalkScanner(rubyWalkEcosystem))
+	reg.Register(newWalkScanner(pythonWalkEcosystem))
+	reg.Register(newWalkScanner(goWalkEcosystem))
 	reg.Register(NewXcodeScanner())
 	reg.Register(NewGlobalScanner())
 	reg.Register(NewLLMScanner())